+++ Src/ImplicitNullability.Plugin/AnnotationRedundancyInHierarchyWarningFilteringDecorator.cs
@@ -1,10 +1,6 @@
 ﻿using System.Collections.Generic;
 using JetBrains.ReSharper.Daemon.CSharp.Errors;
 using JetBrains.ReSharper.Feature.Services.Daemon;
-#if RS20172
-using JetBrains.Application.Settings;
-
-#endif
 
 namespace ImplicitNullability.Plugin
 {
@@ -25,18 +21,11 @@ namespace ImplicitNullability.Plugin
             }
         }
 
-#if RS20172
-        public IDaemonStageProcess Process => _decorated.Process;
-
-        public bool IsNonUserFile => _decorated.IsNonUserFile;
-
-        public bool IsGeneratedFile => _decorated.IsGeneratedFile;
-
-        public IHighlightingSettingsManager HighlightingSettingsManager => _decorated.HighlightingSettingsManager;
-
-        public IContextBoundSettingsStore SettingsStore => _decorated.SettingsStore;
+#if RS20173
+        public IList<HighlightingInfo>Highlightings =>
+#else
+        public IReadOnlyCollection<HighlightingInfo> Highlightings =>
 #endif
-
-        public IList<HighlightingInfo> Highlightings => _decorated.Highlightings;
+            _decorated.Highlightings;
     }
 }