+++ Src/ImplicitNullability.Plugin/ImplicitNullabilityProblemAnalyzer.cs
@@ -12,10 +12,6 @@ using JetBrains.ReSharper.Psi.CSharp.Tree;
 using JetBrains.ReSharper.Psi.Tree;
 using JetBrains.Util;
 
-#if RS20172
-using JetBrains.ReSharper.Daemon.Stages.Dispatcher;
-#endif
-
 namespace ImplicitNullability.Plugin
 {
     [ElementProblemAnalyzer(